pyproject.toml
[tool.poetry]
name = "ml-benchmark"
version = "0.1.0"
description = ""
authors = ["Mawwlle <mawwlle@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scikit-learn = "^1.4.1.post1"
torch = "^2.2.0"
typer = {version="0.9.0", extras=["all"]}
requests = "^2.31.0"
py-cpuinfo = "^9.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
isort = "^5.13.2"
black = "^24.2.0"
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
